Rename package to ottv
