+++ v1/tv/ottv/tracer.go
@@ -1,4 +1,6 @@
-package tracelytics
+// Copyright (C) 2017 Librato, Inc. All rights reserved.
+
+package ottv
 
 import (
 	"sync"
